# openapi.yaml
openapi: 3.1.0
info:
  title: ConstrSW BFF API
  description: |
    Backend for Frontend (BFF) REST API que agrega e simplifica o acesso aos microserviços do sistema ConstrSW.
    
    Esta API atua como uma camada intermediária entre o frontend e os microserviços, fornecendo:
    - Agregação de dados de múltiplos serviços
    - Simplificação de chamadas complexas
    - Redução do número de requisições necessárias
    - Padronização de respostas
    
    ## Autenticação
    Todos os endpoints (exceto `/auth/login` e `/auth/refresh`) requerem autenticação via Bearer Token (JWT).
    O token deve ser obtido através do endpoint `/auth/login` e incluído no header `Authorization: Bearer <token>`.
    
    ## Microserviços Agregados
    - OAuth (Autenticação e Autorização)
    - Employees (Funcionários)
    - Classes (Turmas)
    - Courses (Disciplinas)
    - Lessons (Aulas)
    - Professors (Professores)
    - Reservations (Reservas)
    - Resources (Recursos)
    - Rooms (Salas)
    - Students (Estudantes)
  version: 1.0.0
  contact:
    name: ConstrSW Team
    email: support@constrsw.example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Servidor de desenvolvimento local
  - url: http://api:3000/api/v1
    description: Servidor interno (Docker)
...
paths:
...
  /roles:
    get:
      tags:
        - Roles
      summary: Listar roles
      description: Retorna uma lista de todos os roles cadastrados
      operationId: listRoles
      security:
        - bearerAuth: []
      parameters:
        - name: enabled
          in: query
          description: Filtrar por status (true/false)
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RoleResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Roles
      summary: Criar role
      description: Cria um novo role no sistema
      operationId: createRole
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleCreateRequest'
      responses:
        '201':
          description: Role criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /roles/{roleId}:
    get:
      tags:
        - Roles
      summary: Buscar role por ID
      description: Retorna os detalhes de um role específico
      operationId: getRoleById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoleId'
      responses:
        '200':
          description: Dados do role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Roles
      summary: Atualizar role (completo)
      description: Atualiza todos os dados de um role
      operationId: updateRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: Role atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Roles
      summary: Atualizar role (parcial)
      description: Atualiza parcialmente os dados de um role
      operationId: patchRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoleId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleUpdateRequest'
      responses:
        '200':
          description: Role atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Roles
      summary: Excluir role
      description: Exclui um role do sistema permanentemente
      operationId: deleteRole
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoleId'
      responses:
        '204':
          description: Role excluído com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /features:
    get:
      tags:
        - Resources
      summary: Listar todas as features
      operationId: listFeatures
      security:
        - bearerAuth: []
      parameters:
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de features
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Resources
      summary: Criar uma nova feature
      operationId: createFeature
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureCreateRequest'
      responses:
        '201':
          description: Feature criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /features/{featureId}:
    get:
      tags:
        - Resources
      summary: Buscar feature por ID
      operationId: getFeatureById
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FeatureId'
      responses:
        '200':
          description: Dados da feature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Resources
      summary: Atualizar feature (parcial)
      operationId: patchFeature
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FeatureId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureUpdateRequest'
      responses:
        '200':
          description: Feature atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Resources
      summary: Substituir feature (completo)
      operationId: replaceFeature
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FeatureId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureUpdateRequest'
      responses:
        '200':
          description: Feature substituída com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Resources
      summary: Deletar feature
      operationId: deleteFeature
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FeatureId'
      responses:
        '204':
          description: Feature deletada com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /feature-values/feature/{featureId}:
    get:
      tags:
        - Resources
      summary: Buscar feature values por feature
      operationId: getFeatureValuesByFeature
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/FeatureId'
      responses:
        '200':
          description: Lista de feature values
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FeatureValueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /feature-values/resources/{resourceId}/features/{featureValueId}:
    get:
      tags:
        - Resources
      summary: Buscar feature value específico
      operationId: getResourceFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/FeatureValueId'
      responses:
        '200':
          description: Dados do feature value
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Resources
      summary: Atualizar feature value
      operationId: patchResourceFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/FeatureValueId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureValueUpdateRequest'
      responses:
        '200':
          description: Feature value atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeatureValueResponse'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Resources
      summary: Deletar feature value
      operationId: deleteResourceFeatureValue
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ResourceId'
        - $ref: '#/components/parameters/FeatureValueId'
      responses:
        '204':
          description: Feature value deletado com sucesso
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /rooms/{roomId}/furnitures:
    get:
      tags:
        - Rooms
      summary: Listar mobílias (scaffold - retorna 501)
      operationId: listRoomFurnitures
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '501':
          description: Não implementado
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags:
        - Rooms
      summary: Criar mobília (scaffold - retorna 501)
      operationId: createRoomFurniture
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '501':
          description: Não implementado
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /rooms/{roomId}/furnitures/{furnitureId}:
    get:
      tags:
        - Rooms
      summary: Buscar mobília (scaffold - retorna 501)
      operationId: getRoomFurniture
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - $ref: '#/components/parameters/FurnitureId'
      responses:
        '501':
          description: Não implementado
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags:
        - Rooms
      summary: Substituir mobília (scaffold - retorna 501)
      operationId: updateRoomFurniture
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - $ref: '#/components/parameters/FurnitureId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '501':
          description: Não implementado
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags:
        - Rooms
      summary: Atualizar mobília (scaffold - retorna 501)
      operationId: patchRoomFurniture
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - $ref: '#/components/parameters/FurnitureId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '501':
          description: Não implementado
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags:
        - Rooms
      summary: Deletar mobília (scaffold - retorna 501)
      operationId: deleteRoomFurniture
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/RoomId'
        - $ref: '#/components/parameters/FurnitureId'
      responses:
        '501':
          description: Não implementado
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtido através do endpoint /auth/login
  parameters:
...
    RoleId:
      name: roleId
      in: path
      required: true
      description: ID único do role
      schema:
        type: string
...
    ResourceId:
      name: resourceId
      in: path
      required: true
      description: ID único do recurso (UUID)
      schema:
        type: string
        format: uuid
    RoomId:
      name: roomId
      in: path
      required: true
      description: ID único da sala
      schema:
        type: string
...
    FurnitureId:
      name: furnitureId
      in: path
      required: true
      description: ID único da mobília
      schema:
        type: string
...
    FeatureId:
      name: featureId
      in: path
      required: true
      description: ID único da feature
      schema:
        type: string
    FeatureValueId:
      name: featureValueId
      in: path
      required: true
      description: ID único do feature value
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Mensagem de erro
        status_code:
          type: integer
          description: Código HTTP do erro
...
    RoleResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RoleCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
          default: true
    RoleUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        enabled:
          type: boolean
...
    FeatureResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
          format: uuid
    FeatureCreateRequest:
      type: object
      required:
        - name
        - categoryId
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
          format: uuid
    FeatureUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
          format: uuid
    FeatureValueResponse:
      type: object
      properties:
        id:
          type: string
        value:
          type: string
        featureId:
          type: string
        resourceId:
          type: string
          format: uuid
...
    FeatureValueUpdateRequest:
      type: object
      properties:
        value:
          type: string
        featureId:
          type: string
        resourceId:
          type: string
          format: uuid
  responses:
    BadRequestError:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Não autorizado - Token ausente, inválido ou expirado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'